Signed-off-by: alisanikiforova <52976762+alisanikiforova@users.noreply.github.com>
This commit adds a new FPGA sample that demonstrates the functionality of the Intel(R) FPGA Dynamic Profiler for DPC++ tool.
This tool can be used on FPGA designs to collect profiling information about the design as it runs on FPGA hardware in order to assist with design optimization.
The tutorial includes a README that explains how the tool can be set up and what data it can collect, as well as a simplistic design example that demonstrates how the data it collects can be used to locate areas for optimization.