Legal, functional and consistency considerations.
Intel legal guidelines insist on have copyright on first line. The only exception is when it is impossible.
There also shouldn't be any non-ascii tokens in a text file (the recommendation from legal is to use a lower case '(c)'.
The include guard will work for any compiler (gcc and MSVC alike). This makes the Microsoft-specific pragma redundant.

Also fixed type and case consistency issue.