Implemented exp() and pow2() lookup tables using metaprogramming. Includes creating constexpr functions for pow(float, int) and exp(float). The latter was done using the Taylor series expansion (see mp_math.hpp).
Changes to qfp.hpp to make QFP::ToFP32() constexpr and therefore allow
the LUTs to be constructed at compile time.

Moved all constants to a separate file (constants.hpp) to cleanup main.cpp.

Implemented a base function for creating a constexpr ROM (translates to
an actual ROM in the FPGA). Is the base class for the exp() and pow2() LUTs.

Had to increase he constexpr step length (-fconstexpr-steps=<value>) in the
CMake file to deal with all this fancy metaprogramming.

Some comments and code cleanup.