Bdmoore1 add vscode 0928 (#683)
* Added VS Code to vector-add

* Added terminal instructions

Added a overview of the steps to downloading and building a sample. Steps are focused on using the VS Code extensions.

* Add VS Code instructions

In most of the sample README files, I added instructions for how to learn more about using VS Code. If the sample seemed too complicated for VS Code (i.e., setting conda env or building folder trees), or the sample was not in the oneapi samples browser extension, then I did not include the VS Code instructions.