Relax correctness test for Ch04 matrix multiply
The error accumulates for each row of the matrix, so some implementations do
not pass with an epsilon of 1.0E-06 when using fast math optimizations.

Since the purpose of this sample is to demonstrate usage of SYCL rather than to
teach how to write rigorous tests for matrix multiplications, this commit
simply relaxes the correctness test by using a larger epsilon.