Incorrect number of "../../.." relative references in the commonFolder field for the FPGA "compute_units" and "pipe_array" samples.
This fixes a problem in the sample.json files for these two samples. These
mistakes cannot be found by the sample.json schema checker, so a check for
this condition was added to the sample aggregator's code. The new test
confirms that the files referenced by the commonFolder object do exist.