incorrect "commonFolder" values in two FPGA samples (#928)
* Delete Tools/IoTConnectionTools/analog-in directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/digital-in directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/digital-out directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/hello-iot-world directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/interrupt directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/onboard-blink directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/pwm directory

We are retiring the MRAA/UPM IoT samples.

* Delete Tools/IoTConnectionTools/up2-leds directory

We are retiring the MRAA/UPM IoT samples.

* Incorrect number of "../../.." relative references in the commonFolder
field for the FPGA "compute_units" and "pipe_array" samples.

This fixes a problem in the sample.json files for these two samples. These
mistakes cannot be found by the sample.json schema checker, so a check for
this condition was added to the sample aggregator's code. The new test
confirms that the files referenced by the commonFolder object do exist.