Updated census sample readme
Restructured the readme to match the new template, fixed formatting issues, updated branding, and rewrote it for clarity. Added new information about running it in Intel® DevCloud via JupyterLab, as that seems the most practical method for running this type of sample.

Also updated ..\oneAPI-samples\AI-and-Analytics\End-to-end-Workloads\README.md. Fixed some formatting and branding issues.(Included in this pull request for convenience.)