Duplicate Figure 19-7 as Figure 19-17
The book reproduces the same snippet as both Figure 19-7 and 19-17,
first to introduce the concept of atomics, and second to demonstrate the
usage of atomic_ref with USM.

The fact that 19-17 was "missing" from the repository was reported by
readers as an error, and duplicating the source file may reduce
confusion. We tried adding a symbolic link, but this caused issues on
Windows.