FPGA: Update memory_attributes sample to use IS_BSP flag (#1548)
This change fixes some configuration issues that were causing the -Xsuse-2xclock flag to be wrongly passed in as a compile flag. The main issue was that we do not want to use the flag when doing Full-System compiles. To resolve this we now make use of the IS_BSP flag to distinguish between flows.

This change also updates the default board from Agilex -> Agilex7 to suppress warnings that were showing up in the compiles.