FPGA: Adding matrix multiplication code sample (#1568)
Adding a new general matrix multiplication (GEMM) code sample. Calculates the matrix product using a systolic array. Works with single-precision floating-point matrices and is parameterized on the size of the input matrices and the number of PEs (uses tiling to compute a matrix product larger than the size of the PE array).