FPGA: Remove db query 9 for 2023.2 (#1740)
# Existing Sample Changes
## Description

This change removes db query 9 from the 2023.2 code samples.

## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Implement fixes for ONSAM Jiras

## How Has This Been Tested?
I ran db1, db9, db11, and db12 in emulator flow after this change
https://spetc.intel.com/testsummary?testRunIds=8227959

We see that db9 now has the message that it is not a valid query number