FPGA: Remove throughput metric from loop_coalesce and update README (#2060)
On S10 and Agilex7, the throughput increase described in the README does not appear. In fact, we see a slight decrease in throughput on these boards for some seeds, and the results are equivocal. I changed the sample to no longer report throughput and have the code sample readme reflect this as well.

Instead, I show users the area savings of the loop_coalesce sample, which is more consistent.