FPGA: Merge Sort Sample Quartus Compatibility Fix  (#2086)
# Existing Sample Changes
## Description

The merge_sort reference design fails in the SYCL HLS flow for versions of Quartus which only support up to 32 bit interrupts (22.2 and earlier). The number of kernels generated with `MERGE_UNITS=P` is `5P-2`, thus decreasing it from `P=8` to `P=4` gives us `5P-2 = 22 < 32` as required.

This change introduces a flag LIMIT_HW_MERGE_UNITS that sets the default MERGE_UNITS to 4 when specified for hardware only, and only in SYCL HLS flow (non-BSP targets). This flag is overridden by the MERGE_UNITS flag, and does not affect the MERGE_UNITS value in simulator or emulator mode.

## How Has This Been Tested?
`cmake .. -DFPGA_DEVICE=<family>` with Quartus 21.2 and 23.3
- [x] Command Line