FPGA: Reorder submit order for DB11 and DB12 to get more stable timing/Throughput numbers (#2339)
To minimize the variation on the timing, it is important to start the 'first' kernel of a chain of kernels last. This ensures that the timing doesn't start until that kernel is ready to go.

For DB12, there are 2 pipelines that are joined. If PRECISE_TIMING is defined, add a dummy start kernel that will trigger the two pipelines to start.