FPGA: Update MatrixReadPipeToDDR to use ptr_annotations (#2487)
A recent functional change to the compiler means that it will now correctly identify memory dependences. As a side effect, this will now cause for the compiler to emit a message that it is unable to achieve a user specified II for the MatrixReadPipeToDDR function in memory_transfers.hpp.

To regain this performance we can use annotated_ptr's in the SYCL HLS flow to specify a larger interface width which will allow for the compiler to coalesce stores to memory, thus resulting in being able to achieve the user specified II again.